이인서/2주차/requirements.txt
asgiref==3.5.2
# backports.zoneinfo==0.2.1
dj-rest-auth==2.2.2
Django==4.1
django-js-asset==2.0.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
gunicorn==20.1.0
joblib==1.1.0
numpy==1.23.2
pandas==1.4.3
PyJWT==2.4.0
psycopg2-binary==2.9.3
python-dateutil==2.8.2
pytz==2022.2.1
scikit-learn==1.1.2
scipy==1.9.0
six==1.16.0
sklearn==0.0
sqlparse==0.4.2
threadpoolctl==3.1.0
tzdata==2022.2
# windows-curses==2.3.0
